Refactored the Conosle application to also publish a bootstrapper recipe.
Allows creating a custom console applicatoin that is a Hangfire MessageBus Harness along with ability to control Handler Building, logging, etc. especially useful for later frameworks as well as handlers that have issues loading via reflection.   Also added a pass method that will log a message and a fail method to throw a SimulatedFailureException.